domain step 1
git-svn-id: http://192.168.20.11/sequoiadb/trunk@11553 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsReorgUnit.cpp
@@ -425,7 +425,7 @@ namespace engine
       recordOffset = _currentExtentSize - currentExtent->_freeSpace ;
       recordPtr = ((ossValuePtr)currentExtent) + recordOffset ;
       if ( currentExtent->_freeSpace - (INT32)dmsrecordSize <
-           DMS_MIN_RECORD_SZ &&
+           (INT32)DMS_MIN_RECORD_SZ &&
            currentExtent->_freeSpace <= (INT32)DMS_RECORD_MAX_SZ )
       {
          dmsrecordSize = (UINT32)currentExtent->_freeSpace ;
